Changes some types to be more explicit, also changes the hash to a 32bit integer
git-svn-id: https://svn.cct.lsu.edu/repos/projects/parallex/branches/chess@3813 5079c8e4-9419-0410-93bb-9e5ee278c886

+++ openmp/headers/hash.hpp
@@ -4,7 +4,7 @@
 #include <pthread.h>
 #include <stdio.h>
 
-typedef uint64_t hash_t;
+typedef uint32_t hash_t;
 
 const int zdepth = 2;
 const int N = 8;  // bits for bucket size
